third_party/WebKit/LayoutTests/fast/css/hover-leftmost-quirks-expected.txt

Quirks mode - simple selectors right of :hover

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS window.eventSender is defined.
PASS getComputedStyle(t1).backgroundColor is transparent
PASS getComputedStyle(t1).backgroundColor is green
PASS getComputedStyle(t2).backgroundColor is transparent
PASS getComputedStyle(t2).backgroundColor is green
PASS getComputedStyle(t3).backgroundColor is transparent
PASS getComputedStyle(t3).backgroundColor is green
PASS successfullyParsed is true

TEST COMPLETE

Green background when hovered
Green background when hovered
Green background when hovered
